+++ 3DGameEngine/src/shader.cpp
@@ -243,7 +243,7 @@ void ShaderData::AddProgram(const std::string& text, int type)
 	const GLchar* p[1];
 	p[0] = text.c_str();
 	GLint lengths[1];
-	lengths[0] = text.length();
+	lengths[0] = ((GLint)text.length());
 
 	glShaderSource(shader, 1, p, lengths);
 	glCompileShader(shader);
